[chore] PlayerState 파일 분리
