+++ app/src/main/java/com/squirtles/musicroad/musicplayer/PlayerViewModel.kt
@@ -18,17 +18,6 @@ import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 import javax.inject.Inject
 
-data class PlayerState(
-    val isReady: Boolean = true,
-    val isPlaying: Boolean = false,
-    val currentPosition: Long = 0L,
-) {
-    companion object {
-        val PLAYER_STATE_INITIAL = PlayerState(isReady = false)
-        val PLAYER_STATE_STOP = PlayerState(isReady = true, isPlaying = false)
-    }
-}
-
 @HiltViewModel
 class PlayerViewModel @Inject constructor() : ViewModel() {
